==> 5-eda-pandas.py <==
# ---
# jupyter:
#   jupytext:
#     cell_metadata_filter: -all
#     text_representation:
#       extension: .py
#       format_name: percent
#       format_version: '1.3'
#       jupytext_version: 1.13.8
#   kernelspec:
#     display_name: Python 3 (ipykernel)
#     language: python
#     name: python3
# ---

# %% [markdown]
# <p style="text-align:center">
#     <a href="https://skills.network/?utm_medium=Exinfluencer&utm_source=Exinfluencer&utm_content=000026UJ&utm_term=10006555&utm_id=NA-SkillsNetwork-Channel-SkillsNetworkCoursesIBMDS0321ENSkillsNetwork26802033-2022-01-01" target="_blank">
#     <img src="https://cf-courses-data.s3.us.cloud-object-storage.appdomain.cloud/assets/logos/SN_web_lightmode.png" width="200" alt="Skills Network Logo"  />
#     </a>
# </p>
#

# %% [markdown]
# # **SpaceX  Falcon 9 First Stage Landing Prediction**
#

# %% [markdown]
# ## Assignment: Exploring and Preparing Data
#

# %% [markdown]
# Estimated time needed: **70** minutes
#

# %% [markdown]
# In this assignment, we will predict if the Falcon 9 first stage will land successfully. SpaceX advertises Falcon 9 rocket launches on its website with a cost of 62 million dollars; other providers cost upward of 165 million dollars each, much of the savings is due to the fact that SpaceX can reuse the first stage.
#
# In this lab, you will perform Exploratory Data Analysis and Feature Engineering.
#

# %% [markdown]
# Falcon 9 first stage will land successfully
#

# %% [markdown]
# ![](https://cf-courses-data.s3.us.cloud-object-storage.appdomain.cloud/IBMDeveloperSkillsNetwork-DS0701EN-SkillsNetwork/api/Images/landing\_1.gif)
#

# %% [markdown]
# Several examples of an unsuccessful landing are shown here:
#

# %% [markdown]
# ![](https://cf-courses-data.s3.us.cloud-object-storage.appdomain.cloud/IBMDeveloperSkillsNetwork-DS0701EN-SkillsNetwork/api/Images/crash.gif)
#

# %% [markdown]
# Most unsuccessful landings are planned. Space X performs a controlled landing in the oceans.
#

# %% [markdown]
# ## Objectives
#
# Perform exploratory Data Analysis and Feature Engineering using `Pandas` and `Matplotlib`
#
# *   Exploratory Data Analysis
# *   Preparing Data  Feature Engineering
#

# %% [markdown]
# ***
#

# %% [markdown]
# ### Import Libraries and Define Auxiliary Functions
#

# %% [markdown]
# We will import the following libraries the lab
#

# %%
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# %% [markdown]
# ## Exploratory Data Analysis
#

# %% [markdown]
# First, let's read the SpaceX dataset into a Pandas dataframe and print its summary
#

# %%
df=pd.read_csv("https://cf-courses-data.s3.us.cloud-object-storage.appdomain.cloud/IBM-DS0321EN-SkillsNetwork/datasets/dataset_part_2.csv")

# If you were unable to complete the previous lab correctly you can uncomment and load this csv

# df = pd.read_csv('https://cf-courses-data.s3.us.cloud-object-storage.appdomain.cloud/IBMDeveloperSkillsNetwork-DS0701EN-SkillsNetwork/api/dataset_part_2.csv')

df.head(5)

# %% [markdown]
# First, let's try to see how the `FlightNumber` (indicating the continuous launch attempts.) and `Payload` variables would affect the launch outcome.
#
# We can plot out the <code>FlightNumber</code> vs. <code>PayloadMass</code>and overlay the outcome of the launch. We see that as the flight number increases, the first stage is more likely to land successfully. The payload mass is also important; it seems the more massive the payload, the less likely the first stage will return.
#

# %%
sns.catplot(y="PayloadMass", x="FlightNumber", hue="Class", data=df, aspect = 5)
plt.xlabel("Flight Number",fontsize=20)
plt.ylabel("Pay load Mass (kg)",fontsize=20)
plt.show()

# %% [markdown]
# We see that different launch sites have different success rates.  <code>CCAFS LC-40</code>, has a success rate of 60 %, while  <code>KSC LC-39A</code> and <code>VAFB SLC 4E</code> has a success rate of 77%.
#

# %% [markdown]
# Next, let's drill down to each site visualize its detailed launch records.
#

# %% [markdown]
# ### TASK 1: Visualize the relationship between Flight Number and Launch Site
#

# %% [markdown]
# Use the function <code>catplot</code> to plot <code>FlightNumber</code> vs <code>LaunchSite</code>, set the  parameter <code>x</code>  parameter to <code>FlightNumber</code>,set the  <code>y</code> to <code>Launch Site</code> and set the parameter <code>hue</code> to <code>'class'</code>
#

# %%
sns.catplot(y="LaunchSite", x="FlightNumber", hue="Class", data=df, aspect = 5)
plt.xlabel("Flight Number",fontsize=20)
plt.ylabel("Launch Site",fontsize=20)
plt.show()


# %% [markdown]
# Now try to explain the patterns you found in the Flight Number vs. Launch Site scatter point plots.
#

# %% [markdown]
# ### TASK 2: Visualize the relationship between Payload and Launch Site
#

# %% [markdown]
# We also want to observe if there is any relationship between launch sites and their payload mass.
#

# %%
sns.catplot(y="LaunchSite", x="PayloadMass", hue="Class", data=df, aspect = 5)
plt.xlabel("Pay load Mass (kg)",fontsize=20)
plt.ylabel("Launch Site",fontsize=20)
plt.show()


# %% [markdown]
# Now if you observe Payload Vs. Launch Site scatter point chart you will find for the VAFB-SLC  launchsite there are no  rockets  launched for  heavypayload mass(greater than 10000).
#

# %% [markdown]
# ### TASK  3: Visualize the relationship between success rate of each orbit type
#

# %% [markdown]
# Next, we want to visually check if there are any relationship between success rate and orbit type.
#

# %% [markdown]
# Let's create a `bar chart` for the sucess rate of each orbit
#

# %%
dfg = df.groupby(['Orbit'], as_index=False)['Class'].mean()
sns.barplot(x = 'Orbit', y = 'Class', data=dfg)
plt.show()


# %% [markdown]
# Analyze the ploted bar chart try to find which orbits have high sucess rate.
#

# %% [markdown]
# ### TASK  4: Visualize the relationship between FlightNumber and Orbit type
#

# %% [markdown]
# For each orbit, we want to see if there is any relationship between FlightNumber and Orbit type.
#

# %%
# Plot a scatter point chart with x axis to be FlightNumber and y axis to be the Orbit, and hue to be the class value
sns.catplot(y="Orbit", x="FlightNumber", hue="Class", data=df, aspect = 5)
plt.xlabel("Flight Number",fontsize=20)
plt.ylabel("Orbit",fontsize=20)
plt.show()


# %% [markdown]
# You should see that in the LEO orbit the Success appears related to the number of flights; on the other hand, there seems to be no relationship between flight number when in GTO orbit.
#

# %% [markdown]
# ### TASK  5: Visualize the relationship between Payload and Orbit type
#

# %% [markdown]
# Similarly, we can plot the Payload vs. Orbit scatter point charts to reveal the relationship between Payload and Orbit type
#

# %%
# Plot a scatter point chart with x axis to be Payload and y axis to be the Orbit, and hue to be the class value
sns.catplot(y="Orbit", x="PayloadMass", hue="Class", data=df, aspect = 5)
plt.xlabel("Pay load Mass (kg)",fontsize=20)
plt.ylabel("Launch Site",fontsize=20)
plt.show()


# %% [markdown]
# With heavy payloads the successful landing or positive landing rate are more for Polar,LEO and ISS.
#
# However for GTO we cannot distinguish this well as both positive landing rate and negative landing(unsuccessful mission) are both there here.
#

# %% [markdown]
# ### TASK  6: Visualize the launch success yearly trend
#

# %% [markdown]
# You can plot a line chart with x axis to be <code>Year</code> and y axis to be average success rate, to get the average launch success trend.
#

# %% [markdown]
# The function will help you get the year from the date:
#

# %%
# A function to Extract years from the date 
def extract_year(df):
    year = []
    for i in df["Date"]:
        year.append(i.split("-")[0])
    return pd.DataFrame(year , columns=['Year'])


# %%
# Plot a line chart with x axis to be the extracted year and y axis to be the success rate
df['Year'] = extract_year(df)
df.groupby('Year', as_index=False)['Class'].mean()

sns.lineplot(data=df, x='Year', y='Class')
plt.show()


# %% [markdown]
# you can observe that the sucess rate since 2013 kept increasing till 2020
#

# %% [markdown]
# ## Features Engineering
#

# %% [markdown]
# By now, you should obtain some preliminary insights about how each important variable would affect the success rate, we will select the features that will be used in success prediction in the future module.
#

# %%
features = df[['FlightNumber', 'PayloadMass', 'Flights', 'GridFins', 'Reused', 'Legs', 'Block', 'ReusedCount']]
features.head()

# %% [markdown]
# ### TASK  7: Create dummy variables to categorical columns
#

# %% [markdown]
# Use the function <code>get_dummies</code> and <code>features</code> dataframe to apply OneHotEncoder to the column <code>Orbits</code>, <code>LaunchSite</code>, <code>LandingPad</code>, and <code>Serial</code>. Assign the value to the variable <code>features_one_hot</code>, display the results using the method head. Your result dataframe must include all features including the encoded ones.
#

# %%
# HINT: Use get_dummies() function on the categorical columns


# %%
features = pd.concat([features,pd.get_dummies(df['Orbit'])], axis=1)
features = pd.concat([features,pd.get_dummies(df['LandingPad'])], axis=1)
features = pd.concat([features,pd.get_dummies(df['LaunchSite'])], axis=1)
features = pd.concat([features,pd.get_dummies(df['Serial'])], axis=1)
features.head()

# %% [markdown]
# ### TASK  8: Cast all numeric columns to `float64`
#

# %% [markdown]
# Now that our <code>features_one_hot</code> dataframe only contains numbers cast the entire dataframe to variable type <code>float64</code>
#

# %%
# HINT: use astype function
features = features.astype('float64')


# %%
features.to_csv('dataset_part_3.csv', index=False)

# %% [markdown]
# We can now export it to a <b>CSV</b> for the next section,but to make the answers consistent, in the next lab we will provide data in a pre-selected date range.
#

# %% [markdown]
# <code>features_one_hot.to_csv('dataset_part\_3.csv', index=False)</code>
#

# %% [markdown]
# ## Authors
#

# %% [markdown]
# <a href="https://www.linkedin.com/in/joseph-s-50398b136/?utm_medium=Exinfluencer&utm_source=Exinfluencer&utm_content=000026UJ&utm_term=10006555&utm_id=NA-SkillsNetwork-Channel-SkillsNetworkCoursesIBMDS0321ENSkillsNetwork26802033-2022-01-01">Joseph Santarcangelo</a> has a PhD in Electrical Engineering, his research focused on using machine learning, signal processing, and computer vision to determine how videos impact human cognition. Joseph has been working for IBM since he completed his PhD.
#

# %% [markdown]
# <a href="https://www.linkedin.com/in/nayefaboutayoun/?utm_medium=Exinfluencer&utm_source=Exinfluencer&utm_content=000026UJ&utm_term=10006555&utm_id=NA-SkillsNetwork-Channel-SkillsNetworkCoursesIBMDS0321ENSkillsNetwork26802033-2022-01-01">Nayef Abou Tayoun</a> is a Data Scientist at IBM and pursuing a Master of Management in Artificial intelligence degree at Queen's University.
#

# %% [markdown]
# ## Change Log
#

# %% [markdown]
# | Date (YYYY-MM-DD) | Version | Changed By    | Change Description      |
# | ----------------- | ------- | ------------- | ----------------------- |
# | 2021-10-12        | 1.1     | Lakshmi Holla | Modified markdown       |
# | 2020-09-20        | 1.0     | Joseph        | Modified Multiple Areas |
# | 2020-11-10        | 1.1     | Nayef         | updating the input data |
#

# %% [markdown]
# Copyright © 2020 IBM Corporation. All rights reserved.
#
